fix a bunch of warnings

+++ tools/htf_info.c
@@ -39,7 +39,7 @@ void info_sequence(struct htf_sequence *s) {
 void info_loop(struct htf_loop *l) {
 	printf("{.nb_loops: %d, .token: %x.%x, .nb_iterations: [", l->nb_loops, HTF_TOKEN_TYPE(l->token),
 				 HTF_TOKEN_ID(l->token));
-	DOFOR(i, l->nb_loops - 1) {
+	DOFOR(i, (int)l->nb_loops - 1) {
 		printf("%d, ", l->nb_iterations[i]);
 	}
 	printf("%d]}\n", l->nb_iterations[l->nb_loops - 1]);
